Rename RadioThermostatScheduleJson
