+++ bundles/org.openhab.binding.radiothermostat/src/main/java/org/openhab/binding/radiothermostat/internal/util/RadioThermostatSchedule.java
